Move util function to component and fix dispute timeline icons
