+++ src/components/Disputes/DisputeTimeline.js
@@ -1,6 +1,14 @@
 import React from 'react'
 import styled from 'styled-components'
-import { Accordion, GU, textStyle, useTheme, Timer } from '@aragon/ui'
+import {
+  Accordion,
+  GU,
+  textStyle,
+  useTheme,
+  Timer,
+  IconClose,
+  IconCheck,
+} from '@aragon/ui'
 
 import dayjs from '../../lib/dayjs'
 
@@ -27,9 +35,9 @@ import { dateFormat } from '../../utils/date-utils'
 import { getDisputeTimeLine } from '../../utils/dispute-utils'
 import { numberToWord } from '../../lib/math-utils'
 import {
-  getOutcomeIcon,
   juryOutcomeToString,
   appealRulingToString,
+  OUTCOMES,
 } from '../../utils/crvoting-utils'
 
 const DisputeTimeline = React.memo(function DisputeTimeline({ dispute }) {
@@ -235,6 +243,29 @@ function OutcomeText({ outcome, phase }) {
       )}
     </div>
   )
+}
+
+export function getOutcomeIcon(outcome, theme) {
+  if (!outcome || outcome === OUTCOMES.Refused) {
+    return {
+      Icon: IconClose,
+      color: theme.disabledIcon,
+    }
+  }
+
+  if (outcome === OUTCOMES.Against) {
+    return {
+      Icon: IconClose,
+      color: theme.negative,
+    }
+  }
+
+  if (outcome === OUTCOMES.InFavor) {
+    return {
+      Icon: IconCheck,
+      color: theme.positive,
+    }
+  }
 }
 
 function PhaseIcon({ phase, active }) {
